Extracting AwsConnectionSettingsSelector out of AwsConnectionSettingsEditor to facilitate AWS connection selection independently of Intellij settings configuration
